refactor(transport-sim): inline trivial locals in random_op and handle_echo_delayed
Drop the one-use roll binding, rename unused loop counter to _, and
fold the delay_ms cast next to its sim_random_range source.

+++ moonpool-transport-sim/src/process.rs
@@ -96,14 +96,12 @@ fn handle_echo_delayed(
     };
 
     let time = ctx.providers().time().clone();
-    let delay_ms = moonpool_sim::sim_random_range(1..100);
+    let delay_ms = moonpool_sim::sim_random_range(1..100) as u64;
     drop(
         ctx.providers()
             .task()
             .spawn_task("echo_delayed_reply", async move {
-                let _ = time
-                    .sleep(std::time::Duration::from_millis(delay_ms as u64))
-                    .await;
+                let _ = time.sleep(std::time::Duration::from_millis(delay_ms)).await;
                 assert_sometimes!(true, "echo_delayed_request_handled");
                 reply.send(response);
             }),
